[tasks] Change default task max_retries to infinite.
The max_retries parameter, if not set, now defaults to None,
which is now interpreted by the scheduler as not having a
retry limit.

This is a workaround to prevent tasks from permanently disabling
if, say, a backend service has an extended outage.  It's expected
to not be necessary after issue #76 is resolved and after delays
start working as described in this issue comment for SirMordred:

https://github.com/chaoss/grimoirelab-sirmordred/issues/347#issuecomment-540962247